Force `Content-Type: application/octet-stream` on media
Thus, media-files are often downloaded directly and not displayed in the browser.
As media-files are user-supplied, this is seems as an feature.

The only exception is the linkmap-folder, where Inyoka will place all icons
and auto-generated CSS for the linkmap. Especially, the CSS would be rejected
by browsers, if the generic content type would be set.