Use non-root user for inyoka
In the inyoka-container the user 'inyoka' is created during build. It hase the uid 998
and gid 998 (the redis-container f.e. uses uid and gid 999).
For the uid/gid range of system users, refer to
<https://manpages.ubuntu.com/manpages/focal/en/man8/useradd.8.html>:

       SYS_UID_MAX (number), SYS_UID_MIN (number)
           Range of user IDs used for the creation of system users by useradd or newusers.

           The default value for SYS_UID_MIN (resp.  SYS_UID_MAX) is 101 (resp.  UID_MIN-1).

       UID_MAX (number), UID_MIN (number)
           Range of user IDs used for the creation of regular users by useradd or newusers.

           The default value for UID_MIN (resp.  UID_MAX) is 1000 (resp. 60000).:uµ

The volumes & and some must-have write-able files are owned by inyoka.
All other files (like inyoka python-sources) are read-only, as they are owned by root.