Add custom error page
Only one template html is provided that is used for every error
(note the `{{placeholder` in the HTML).

This will render f.e. a custom 502 page, if the inyoka-worker is not alive.
In comparison, something like the following snippet will rewrite errors
from the inyoka-worker to a custom error page:

```
reverse_proxy inyoka-worker:8000 {
       @error status 500 502 503 504
       handle_response @error {
               root    * /srv/www/templates/
               rewrite * /error_template.html
               templates
               file_server

               header Cache-Control no-cache
               header Connection close
       }
}
```