Include only production dependencies in container image
The development dependencies are installed via the development swarm file.
The installation has to be done as root, as it's done in the image (the inyoka user
can not change the dependencies. That is a wanted effect for production).

After the dependencies are installed, the current local user is used to
execute the service command. That makes it possible to access the bind-mounts.