Enable snapshotting for redis
This is f.e. useful, if the server machine is rebooted. With this change,
a previous snapshot is imported after the reboot. The data in the snapshot
can be directly used without recalculation. The cache snapshot has not to
be super fresh.
The save durations are taken from the old ubuntuusers-setup.

The redis dumps are saved in an own volume. It is optional to backup, as
all contents in redis can be regenerated.